requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# API et Déploiement
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0

# Data Science et Visualisation
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter et Développement
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.5.0

# Utilitaires
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Kaggle (pour télécharger les datasets)
kaggle>=1.5.0

# Métriques et Evaluation
torchmetrics>=1.0.0

# Configuration et Logging
hydra-core>=1.3.0
wandb>=0.15.0  # optionnel pour tracking